requirements.txt
python-telegram-bot==21.10
selenium==3.141.0
